SUBAPI-78: Broke Down API Abstractions
