+++ Submarine API/Api.Abstractions.UnitTests/Api.Abstractions.UnitTests.csproj
@@ -12,6 +12,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\Api.Abstractions.Authentication\Api.Abstractions.Authentication.csproj" />
       <ProjectReference Include="..\Api.Abstractions\Api.Abstractions.csproj" />
     </ItemGroup>
