+++ Submarine API/Api/Api.csproj
@@ -11,8 +11,10 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\Api.Abstractions.Authentication\Api.Abstractions.Authentication.csproj" />
       <ProjectReference Include="..\..\Submarine Domain Authentication\Domain.Authentication\Domain.Authentication.csproj" />
       <ProjectReference Include="..\..\Submarine Domain\Domain\Domain.csproj" />
+      <ProjectReference Include="..\Api.Abstractions.Swagger\Api.Abstractions.Swagger.csproj" />
       <ProjectReference Include="..\Api.Abstractions\Api.Abstractions.csproj" />
     </ItemGroup>
