+++ Submarine API/Api/Controllers/UserController.cs
@@ -2,7 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using Diagnosea.Submarine.Abstractions.Enums;
-using Diagnosea.Submarine.Api.Abstractions.Attributes;
+using Diagnosea.Submarine.Api.Abstractions.Authentication.Attributes;
 using Diagnosea.Submarine.Api.Abstractions.Extensions;
 using Diagnosea.Submarine.Domain.Instructors.User;
 using Microsoft.AspNetCore.Http;
