+++ Submarine API/Api/Startup.cs
@@ -1,4 +1,6 @@
+using Diagnosea.Submarine.Api.Abstractions.Authentication.Extensions;
 using Diagnosea.Submarine.Api.Abstractions.Extensions;
+using Diagnosea.Submarine.Api.Abstractions.Swagger.Extensions;
 using Diagnosea.Submarine.Api.Settings;
 using Diagnosea.Submarine.Domain.Authentication.Extensions;
 using Diagnosea.Submarine.Domain.Extensions;
